Use double to avoid truncating loc timestamp

+++ Footprints/Core/GPS/GPSLocatable.swift
@@ -11,7 +11,7 @@ protocol GPSLocatable: Timestamped {
     var latitude: CGFloat { get }
     var longitude: CGFloat { get }
     var altitude: Measurement<UnitLength> { get }
-    var timestamp: Float { get }
+    var timestamp: Double { get }
     var speed: Double { get }
     func distance(from loc: GPSLocatable) -> Measurement<UnitLength>
 }
